refactor(inputcheckbox): created Inputs folder in storybook for inputs
