+++ src/core/InputCheckbox/index.test.tsx
