+++ src/core/InputCheckbox/index.tsx
