+++ src/core/InputCheckbox/styles.ts
