+++ src/index.ts
@@ -22,8 +22,6 @@ export * from "./core/CellComponent";
 export { default as CellComponent } from "./core/CellComponent";
 export * from "./core/CellHeader";
 export { default as CellHeader } from "./core/CellHeader";
-export * from "./core/Checkbox";
-export { default as Checkbox } from "./core/Checkbox";
 export * from "./core/Chip";
 export { default as Chip } from "./core/Chip";
 export * from "./core/ComplexFilter";
@@ -44,6 +42,8 @@ export * from "./core/DropdownMenu";
 export { default as DropdownMenu } from "./core/DropdownMenu";
 export * from "./core/Icon";
 export { default as Icon } from "./core/Icon";
+export * from "./core/InputCheckbox";
+export { default as Checkbox } from "./core/InputCheckbox";
 export * from "./core/InputDropdown";
 export { default as InputDropdown } from "./core/InputDropdown";
 export * from "./core/InputSearch";
